remove extra routes
