+++ renderer/components/banner/Banner.module.css
@@ -28,6 +28,7 @@
   font-size: 1.5rem;
 }
 .Banner h1 {
+  margin-top: 3rem;
   flex: 4;
   font-size: 8vh;
   line-height: 3vh;
